+++ source/Tubeshade.Data/Migrations/000041_feed-update-type.sql
@@ -1,0 +1,1 @@
+﻿ALTER TYPE tasks.task_type ADD VALUE 'youtube_feed_update';