+++ source/Tubeshade.Data/Migrations/000042_task-payload.sql
@@ -1,0 +1,2 @@
+﻿ALTER TABLE tasks.tasks
+    ADD COLUMN payload TEXT NULL;