+++ source/Tubeshade.Data/Tasks/TaskType.cs
@@ -13,6 +13,7 @@ public sealed class TaskType : SmartEnum<TaskType>
     public static readonly TaskType RefreshSubscriptions = new(Names.RefreshSubscriptions, 6);
     public static readonly TaskType ReindexVideos = new(Names.ReindexVideos, 7);
     public static readonly TaskType UpdateSponsorBlockSegments = new(Names.UpdateSponsorBlockSegments, 8);
+    public static readonly TaskType YouTubeFeedUpdate = new(Names.YouTubeFeedUpdate, 9);
 
     public bool IsLibraryTask =>
         Equals(ScanSubscriptions) ||
@@ -37,5 +38,6 @@ public sealed class TaskType : SmartEnum<TaskType>
         public const string RefreshSubscriptions = "refresh_subscriptions";
         public const string ReindexVideos = "reindex_videos";
         public const string UpdateSponsorBlockSegments = "update_sponsor_block_segments";
+        public const string YouTubeFeedUpdate = "youtube_feed_update";
     }
 }